requirements.txt
customtkinter==5.2.1
Pillow==10.0.1
Pillow==10.1.0
SQLAlchemy==1.3.10
tkcalendar==1.6.1
